Fixed StorageObject deletion race.
* If StorageObject status update is too slow, we might go beyond the end of the
  attached LUNs list raising a StopIteration exception. Refactored to avoid
  that.
* Removed a couple of adjacent obsolete FIXMEs

+++ rtslib/tcm.py
@@ -441,12 +441,11 @@ class StorageObject(CFSNode):
 
         # If we are called after a configure error, we can skip this
         if self.is_configured():
-            gen = self._gen_attached_luns()
-            while self.status == 'activated':
-                gen.next().delete()
-
-            # FIXME: delete ALUA tp pt gps
-            # FIXME: delete snapshots
+            for lun in self._gen_attached_luns():
+                if self.status != 'activated':
+                    break
+                else:
+                    gen.next().delete()
 
         super(StorageObject, self).delete()
